refactor: frame 패키지 생성
