+++ src/main/java/bowling/domain/engine/frame/Frame.java
@@ -1,5 +1,6 @@
-package bowling.domain.engine;
+package bowling.domain.engine.frame;
 
+import bowling.domain.engine.PitchResult;
 import bowling.dto.RecordsDto;
 
 public interface Frame {
